@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -46,6 +46,9 @@ dependencies:
   html: ^0.15.0
   pin_code_fields: ^7.3.0
   flutter_spinkit: ^5.1.0
+  flutter_secure_storage: ^4.2.1
+  localstorage: ^4.0.0+1
+  modal_progress_hud_nsn: ^0.1.0-nullsafety-1
 
 
 dev_dependencies:
