@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:vixx_app/Config/Color.dart';
 import 'package:vixx_app/Config/Style.dart';
 import 'package:vixx_app/Screen/Account/account_info_page.dart';
@@ -10,22 +11,53 @@ import 'package:vixx_app/Screen/Pages/deposit_page.dart';
 import 'package:vixx_app/Screen/Pages/invest_page.dart';
 import 'package:vixx_app/Screen/Pages/transaction_history.dart';
 import 'package:vixx_app/Screen/Pages/withdrawal_page.dart';
+import 'package:vixx_app/Screen/login.dart';
 
 class HomeScreen extends StatefulWidget{
+
+
   @override
   HomeScreenState createState() => HomeScreenState();
 }
 
 
-class HomeScreenState extends State  <HomeScreen>{
+class HomeScreenState extends State  <HomeScreen> {
 
-  String totalAmount = "0", totalEarnings = "0", accumulatedBal= "0";
+  SharedPreferences? sharedPreferences;
 
+  String totalAmount = "0",
+      totalEarnings = "0",
+      accumulatedBal = "0";
+
+
+
+
+
+
+    getData() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    if (prefs.getString("lastname",) == null)
+      setState(() => lastname = '0',);
+    else
+      setState(() => lastname = prefs.getString("lastname")!);
+  }
+
+  getfirstname() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    if (prefs.getString("firstname",) == null)
+      setState(() => firstname = '0',);
+    else
+      setState(() => firstname = prefs.getString("firstname")!);
+  }
+String firstname = '0';
+String lastname = '0';
   @override
   void initState() {
     // TODO: implement initState
     displaySummary();
     super.initState();
+    getData();
+    getfirstname();
   }
 
   @override
@@ -39,7 +71,10 @@ class HomeScreenState extends State  <HomeScreen>{
     // TODO: implement didChangeDependencies
     super.didChangeDependencies();
   }
-  String data;
+
+
+  String? data;
+
   @override
   Widget build(BuildContext context) {
     // TODO: implement build
@@ -59,22 +94,25 @@ class HomeScreenState extends State  <HomeScreen>{
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 Row(
-                  mainAxisAlignment:  MainAxisAlignment.spaceBetween,
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
-                    Text("Hello", style: setStyleContent(context,Colors.black,15,FontWeight.normal)),
+                    Text("Hello", style: setStyleContent(
+                        context, Colors.black, 15, FontWeight.normal)),
                     GestureDetector(
-                      child:  Icon(Icons.person_outline_sharp),
-                      onTap: (){
-                        Navigator.push(context,MaterialPageRoute(builder: (context) => AccountInformation()));
+                      child: Icon(Icons.person_outline_sharp),
+                      onTap: () {
+                        Navigator.push(context, MaterialPageRoute(builder: (
+                            context) => AccountInformation()));
                       },
                     ),
                   ],
                 ),
-                Text("Elsie Roberts", style: setStyleContent(context,ColorConstant.primaryColor,25,FontWeight.bold)),
+                Text("$firstname $lastname", style: setStyleContent(
+                    context, ColorConstant.primaryColor, 15, FontWeight.bold)),
                 SizedBox(height: 10,),
                 //balance detail page
                 GestureDetector(
-                  child:  Container(
+                  child: Container(
                     padding: EdgeInsets.all(25.0),
                     decoration: BoxDecoration(
                       color: ColorConstant.primaryColor,
@@ -83,21 +121,35 @@ class HomeScreenState extends State  <HomeScreen>{
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
-                        Text("Balance", style: setStyleContent(context,ColorConstant.slightWhiteColor,12,FontWeight.w300)),
-                        Text("USDT ${totalAmount}", style: setStyleContent(context,ColorConstant.slightWhiteColor,25,FontWeight.bold)),
+                        Text("Balance", style: setStyleContent(context,
+                            ColorConstant.slightWhiteColor, 12,
+                            FontWeight.w300)),
+                        Text("USDT ${totalAmount}", style: setStyleContent(
+                            context, ColorConstant.slightWhiteColor, 25,
+                            FontWeight.bold)),
                         SizedBox(height: 11.0,),
                         Row(
                           mainAxisAlignment: MainAxisAlignment.spaceBetween,
                           children: [
-                            Text("Total Credit", style: setStyleContent(context,ColorConstant.slightWhiteColor,12,FontWeight.w300)),
-                            Text("Total Debit", style: setStyleContent(context,ColorConstant.slightWhiteColor,12,FontWeight.w300)),
+                            Text("Total Credit", style: setStyleContent(
+                                context, ColorConstant.slightWhiteColor, 12,
+                                FontWeight.w300)),
+                            Text("Total Debit", style: setStyleContent(
+                                context, ColorConstant.slightWhiteColor, 12,
+                                FontWeight.w300)),
                           ],
                         ),
                         Row(
                           mainAxisAlignment: MainAxisAlignment.spaceBetween,
                           children: [
-                            Text("USDT ${totalEarnings}", style: setStyleContent(context,ColorConstant.slightWhiteColor,20,FontWeight.bold)),
-                            Text("USDT ${accumulatedBal}", style: setStyleContent(context,ColorConstant.slightWhiteColor,20,FontWeight.bold)),
+                            Text("USDT ${totalEarnings}",
+                                style: setStyleContent(
+                                    context, ColorConstant.slightWhiteColor, 20,
+                                    FontWeight.bold)),
+                            Text("USDT ${accumulatedBal}",
+                                style: setStyleContent(
+                                    context, ColorConstant.slightWhiteColor, 20,
+                                    FontWeight.bold)),
                           ],
                         )
                       ],
@@ -122,30 +174,43 @@ class HomeScreenState extends State  <HomeScreen>{
                           Row(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Image.asset("assets/images/green_rectangle.png", height: 20,),
+                              Image.asset("assets/images/green_rectangle.png",
+                                height: 20,),
                               SizedBox(width: 10,),
                               Column(
-                                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                                mainAxisAlignment: MainAxisAlignment
+                                    .spaceBetween,
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: [
-                                  Text("Income", style: setStyleContent(context,ColorConstant.black,14,FontWeight.w300),),
-                                  Text("+ \$20,402.00", style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.bold)),
+                                  Text("Income", style: setStyleContent(
+                                      context, ColorConstant.black, 14,
+                                      FontWeight.w300),),
+                                  Text("usdt ${totalEarnings}", style: setStyleContent(
+                                      context, ColorConstant.greenColor, 15,
+                                      FontWeight.bold)),
                                 ],
                               )
                             ],
                           ),
-                          Image.asset("assets/images/div.png", width: 1, color: ColorConstant.primaryColor,),
+                          Image.asset("assets/images/div.png", width: 1,
+                            color: ColorConstant.primaryColor,),
                           Row(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Image.asset("assets/images/red_rectangle.png", height: 20,),
+                              Image.asset(
+                                "assets/images/red_rectangle.png", height: 20,),
                               SizedBox(width: 10,),
                               Column(
-                                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                                mainAxisAlignment: MainAxisAlignment
+                                    .spaceBetween,
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: [
-                                  Text("Expenses", style: setStyleContent(context,ColorConstant.black,14,FontWeight.w300),),
-                                  Text("- \$10,400.04", style: setStyleContent(context,ColorConstant.pinkColor,15,FontWeight.bold)),
+                                  Text("Expenses", style: setStyleContent(
+                                      context, ColorConstant.black, 14,
+                                      FontWeight.w300),),
+                                  Text("usdt ${accumulatedBal}", style: setStyleContent(
+                                      context, ColorConstant.pinkColor, 15,
+                                      FontWeight.bold)),
                                 ],
                               )
                             ],
@@ -162,7 +227,7 @@ class HomeScreenState extends State  <HomeScreen>{
                     scrollDirection: Axis.horizontal,
                     children: <Widget>[
                       GestureDetector(
-                        child:  Container(
+                        child: Container(
                           width: 130.0,
                           padding: EdgeInsets.all(12),
                           decoration: BoxDecoration(
@@ -172,14 +237,21 @@ class HomeScreenState extends State  <HomeScreen>{
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Image.asset("assets/images/in_icon.png", height: 35,),
-                              Text("Invest", style: setStyleContent(context,ColorConstant.slightWhiteColor,15,FontWeight.bold)),
-                              Text("Invest from as\nlow as 20USDT", style: setStyleContent(context,ColorConstant.slightWhiteColor,10,FontWeight.normal)),
+                              Image.asset(
+                                "assets/images/in_icon.png", height: 35,),
+                              Text("Invest", style: setStyleContent(
+                                  context, ColorConstant.slightWhiteColor, 15,
+                                  FontWeight.bold)),
+                              Text("Invest from as\nlow as 20USDT",
+                                  style: setStyleContent(
+                                      context, ColorConstant.slightWhiteColor,
+                                      10, FontWeight.normal)),
                             ],
                           ),
                         ),
-                        onTap: (){
-                          Navigator.push(context,MaterialPageRoute(builder: (context) => InvestPage()));
+                        onTap: () {
+                          Navigator.push(context, MaterialPageRoute(
+                              builder: (context) => InvestPage()));
                         },
                       ),
                       SizedBox(width: 7,),
@@ -194,19 +266,26 @@ class HomeScreenState extends State  <HomeScreen>{
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Image.asset("assets/images/in_icon.png", height: 35,),
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.slightWhiteColor,15,FontWeight.bold)),
-                              Text("Deposit funds as\nlow as 20USDT", style: setStyleContent(context,ColorConstant.slightWhiteColor,10,FontWeight.normal)),
+                              Image.asset(
+                                "assets/images/in_icon.png", height: 35,),
+                              Text("Deposit", style: setStyleContent(
+                                  context, ColorConstant.slightWhiteColor, 15,
+                                  FontWeight.bold)),
+                              Text("Deposit funds as\nlow as 20USDT",
+                                  style: setStyleContent(
+                                      context, ColorConstant.slightWhiteColor,
+                                      10, FontWeight.normal)),
                             ],
                           ),
                         ),
-                        onTap: (){
-                          Navigator.push(context,MaterialPageRoute(builder: (context) => DepositPage()));
+                        onTap: () {
+                          Navigator.push(context, MaterialPageRoute(
+                              builder: (context) => DepositPage()));
                         },
                       ),
                       SizedBox(width: 7,),
                       GestureDetector(
-                        child:  Container(
+                        child: Container(
                           width: 130.0,
                           padding: EdgeInsets.all(12),
                           decoration: BoxDecoration(
@@ -216,19 +295,26 @@ class HomeScreenState extends State  <HomeScreen>{
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Image.asset("assets/images/with_icon.png", height: 35,),
-                              Text("Withdraw", style: setStyleContent(context,ColorConstant.slightWhiteColor,15,FontWeight.bold)),
-                              Text("Withdraw funds as\nlow as 20USDT", style: setStyleContent(context,ColorConstant.slightWhiteColor,10,FontWeight.normal)),
+                              Image.asset(
+                                "assets/images/with_icon.png", height: 35,),
+                              Text("Withdraw", style: setStyleContent(
+                                  context, ColorConstant.slightWhiteColor, 15,
+                                  FontWeight.bold)),
+                              Text("Withdraw funds as\nlow as 20USDT",
+                                  style: setStyleContent(
+                                      context, ColorConstant.slightWhiteColor,
+                                      10, FontWeight.normal)),
                             ],
                           ),
                         ),
-                        onTap: (){
-                          Navigator.push(context,MaterialPageRoute(builder: (context) => WithdrawalPage()));
+                        onTap: () {
+                          Navigator.push(context, MaterialPageRoute(
+                              builder: (context) => WithdrawalPage()));
                         },
                       ),
                       SizedBox(width: 7,),
                       GestureDetector(
-                        child:  Container(
+                        child: Container(
                           width: 130.0,
                           padding: EdgeInsets.all(12),
                           decoration: BoxDecoration(
@@ -238,13 +324,18 @@ class HomeScreenState extends State  <HomeScreen>{
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Image.asset("assets/images/tran_history.png", height: 35,),
-                              Text("Transaction\nHistory", style: setStyleContent(context,ColorConstant.primaryColor,15,FontWeight.bold)),
+                              Image.asset(
+                                "assets/images/tran_history.png", height: 35,),
+                              Text("Transaction\nHistory",
+                                  style: setStyleContent(
+                                      context, ColorConstant.primaryColor, 15,
+                                      FontWeight.bold)),
                             ],
                           ),
                         ),
-                        onTap: (){
-                          Navigator.push(context,MaterialPageRoute(builder: (context) => TransactionHistoryPage()));
+                        onTap: () {
+                          Navigator.push(context, MaterialPageRoute(
+                              builder: (context) => TransactionHistoryPage()));
                         },
                       ),
                     ],
@@ -256,14 +347,18 @@ class HomeScreenState extends State  <HomeScreen>{
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
-                        Text("Transaction", style: setStyleContent(context,ColorConstant.black,15,FontWeight.w700)),
-                       GestureDetector(
-                         child:  Icon(Icons.arrow_forward_ios_sharp, size: 20,),
-                         onTap: (){
-
-                           Navigator.push(context,MaterialPageRoute(builder: (context) => TransactionHistoryPage()));
-                         },
-                       )
+                        Text("Transaction", style: setStyleContent(
+                            context, ColorConstant.black, 15, FontWeight.w700)),
+                        GestureDetector(
+                          child: Text("View all", style: setStyleContent(
+                              context, ColorConstant.black, 13,
+                              FontWeight.w600)),
+                          onTap: () {
+                            Navigator.push(context, MaterialPageRoute(
+                                builder: (context) =>
+                                    TransactionHistoryPage()));
+                          },
+                        )
                       ],
                     ),
                   ],
@@ -272,7 +367,7 @@ class HomeScreenState extends State  <HomeScreen>{
                 Column(
                   children: [
                     GestureDetector(
-                      child:  Row(
+                      child: Row(
                         children: [
                           Container(
                             padding: EdgeInsets.all(10),
@@ -283,25 +378,42 @@ class HomeScreenState extends State  <HomeScreen>{
                                 color: ColorConstant.pinkColor
                             ),
                             child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
+                              child: Image.asset(
+                                "assets/images/wallet_icon.png", height: 25,),
                             ),
                           ),
                           SizedBox(width: 12,),
                           Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Text("Withdrawal", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("BTC", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w400)),
+                              Text("USDT Withdrawal", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.w400)),
+                              Text("Successful", style: setStyleContent(
+                                  context, ColorConstant.greenColor, 10,
+                                  FontWeight.w400)),
                             ],
                           ),
                           Spacer(),
-                          Text("\$3,019.00", style: setStyleContent(context,ColorConstant.black,15,FontWeight.w700))
+                          Column(
+                            children: [
+                              Text("\$3,019.00", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.bold)),
+                              Text("2 days ago", style: setStyleContent(
+                                  context, ColorConstant.black, 10,
+                                  FontWeight.w300))
+                            ],
+                          )
                         ],
                       ),
                     ),
-                    SizedBox(height: 12,),
-                    Divider(thickness: 3,),
-                    SizedBox(height: 12,),
+                    SizedBox(height: 10,),
+                    Divider(
+                      thickness: 2,
+                      indent: 50,
+                    ),
+                    SizedBox(height: 10,),
                     GestureDetector(
                       child: Row(
                         children: [
@@ -314,19 +426,220 @@ class HomeScreenState extends State  <HomeScreen>{
                                 color: ColorConstant.greenColor
                             ),
                             child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
+                              child: Image.asset(
+                                "assets/images/wallet_icon.png", height: 25,),
                             ),
                           ),
                           SizedBox(width: 12,),
                           Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Text("Investment", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("BTC", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w400)),
+                              Text("USDT Withdrawal", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.w400)),
+                              Text("Failed", style: setStyleContent(
+                                  context, ColorConstant.pinkColor, 10,
+                                  FontWeight.w400)),
                             ],
                           ),
                           Spacer(),
-                          Text("\$3,019.00", style: setStyleContent(context,ColorConstant.black,15,FontWeight.w700))
+                          Column(
+                            children: [
+                              Text("\$7,038", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.bold)),
+                              Text("Sep 27, 2019", style: setStyleContent(
+                                  context, ColorConstant.black, 10,
+                                  FontWeight.w300))
+                            ],
+                          )
+                        ],
+                      ),
+                    ),
+                    SizedBox(height: 10,),
+                    Divider(
+                      thickness: 2,
+                      indent: 50,
+                    ),
+                    GestureDetector(
+                      child: Row(
+                        children: [
+                          Container(
+                            padding: EdgeInsets.all(10),
+                            height: 40,
+                            width: 40,
+                            decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(8),
+                                color: ColorConstant.pinkColor
+                            ),
+                            child: Center(
+                              child: Image.asset(
+                                "assets/images/wallet_icon.png", height: 25,),
+                            ),
+                          ),
+                          SizedBox(width: 12,),
+                          Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text("BTC Withdrawal", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.w400)),
+                              Text("Failed", style: setStyleContent(
+                                  context, ColorConstant.pinkColor, 10,
+                                  FontWeight.w400)),
+                            ],
+                          ),
+                          Spacer(),
+                          Column(
+                            children: [
+                              Text("\$60,337", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.bold)),
+                              Text("Oct 01, 2019", style: setStyleContent(
+                                  context, ColorConstant.black, 10,
+                                  FontWeight.w300))
+                            ],
+                          )
+                        ],
+                      ),
+                    ),
+                    SizedBox(height: 10,),
+                    Divider(
+                      thickness: 2,
+                      indent: 50,),
+                    GestureDetector(
+                      child: Row(
+                        children: [
+                          Container(
+                            padding: EdgeInsets.all(10),
+                            height: 40,
+                            width: 40,
+                            decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(8),
+                                color: ColorConstant.greenColor
+                            ),
+                            child: Center(
+                              child: Image.asset(
+                                "assets/images/wallet_icon.png", height: 25,),
+                            ),
+                          ),
+                          SizedBox(width: 12,),
+                          Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text("BTC Withdrawal", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.w400)),
+                              Text("Failed", style: setStyleContent(
+                                  context, ColorConstant.pinkColor, 10,
+                                  FontWeight.w400)),
+                            ],
+                          ),
+                          Spacer(),
+                          Column(
+                            children: [
+                              Text("\$60,337", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.bold)),
+                              Text("Oct 01, 2019", style: setStyleContent(
+                                  context, ColorConstant.black, 10,
+                                  FontWeight.w300))
+                            ],
+                          )
+                        ],
+                      ),
+                    ),
+                    SizedBox(height: 10,),
+                    Divider(
+                      thickness: 2,
+                      indent: 50,
+                    ),
+                    GestureDetector(
+                      child: Row(
+                        children: [
+                          Container(
+                            padding: EdgeInsets.all(10),
+                            height: 40,
+                            width: 40,
+                            decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(8),
+                                color: ColorConstant.pinkColor
+                            ),
+                            child: Center(
+                              child: Image.asset(
+                                "assets/images/wallet_icon.png", height: 25,),
+                            ),
+                          ),
+                          SizedBox(width: 12,),
+                          Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text("BTC Withdrawal", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.w400)),
+                              Text("Successful", style: setStyleContent(
+                                  context, ColorConstant.greenColor, 10,
+                                  FontWeight.w400)),
+                            ],
+                          ),
+                          Spacer(),
+                          Column(
+                            children: [
+                              Text("\$60,337", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.bold)),
+                              Text("Oct 01, 2019", style: setStyleContent(
+                                  context, ColorConstant.black, 10,
+                                  FontWeight.w300))
+                            ],
+                          )
+                        ],
+                      ),
+                    ),
+                    SizedBox(height: 10,),
+                    Divider(
+                      thickness: 2,
+                      indent: 50,
+                    ),
+                    GestureDetector(
+                      child: Row(
+                        children: [
+                          Container(
+                            padding: EdgeInsets.all(10),
+                            height: 40,
+                            width: 40,
+                            decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(8),
+                                color: ColorConstant.greenColor
+                            ),
+                            child: Center(
+                              child: Image.asset(
+                                "assets/images/wallet_icon.png", height: 25,),
+                            ),
+                          ),
+                          SizedBox(width: 12,),
+                          Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text("USDT Withdrawal", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.w400)),
+                              Text("Successful", style: setStyleContent(
+                                  context, ColorConstant.greenColor, 10,
+                                  FontWeight.w400)),
+                            ],
+                          ),
+                          Spacer(),
+                          Column(
+                            children: [
+                              Text("\$60,337", style: setStyleContent(
+                                  context, ColorConstant.black, 12,
+                                  FontWeight.bold)),
+                              Text("Oct 15, 2019", style: setStyleContent(
+                                  context, ColorConstant.black, 10,
+                                  FontWeight.w300))
+                            ],
+                          )
                         ],
                       ),
                     ),
@@ -339,18 +652,23 @@ class HomeScreenState extends State  <HomeScreen>{
       ),
     );
   }
+
   Future<void> displaySummary() async {
     Map data = {
       'type': null,
     };
     print(data);
+
     String body = json.encode(data);
+    var jsonResponse = null;
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    String? apikey = prefs.getString("apikey");
     String Url = "https://thevix.club/apimaster.php";
     var response = await http.post(Uri.parse(Url),
         headers: {
           "Content-Type": "application/json",
           "Operation": "summary",
-          "Authorization": "hthwijlewivcbeusnwjx6yqldbdi"
+          "Authorization": "$apikey"
         },
 
 
@@ -358,18 +676,31 @@ class HomeScreenState extends State  <HomeScreen>{
         encoding: Encoding.getByName("utf-8")
     );
     print(response.statusCode);
-    if (response.statusCode == 200) {
-      print(response.body.toString()); // for Printing the token
-      var data = response.body;
-      var  userAccount = json.decode(data);
+    if (response.statusCode == 200)
+      jsonResponse = json.decode(response.body);
+    print(response.body.toString()); // for Printing the token
+    if (jsonResponse['status'] == 'success') {
       setState(() {
-        totalAmount = userAccount['data']['balance'];
-        totalEarnings = userAccount['data']['total_credit'];
-        accumulatedBal = userAccount['data']['total_debit'];
+        totalAmount = jsonResponse['data']['balance'];
+        totalEarnings = jsonResponse['data']['total_credit'];
+        accumulatedBal = jsonResponse['data']['total_debit'];
       });
-    } else {
+    }
+    else {
       print('error');
     }
   }
-}
 
+  // getData() async{
+  //   SharedPreferences prefs = await SharedPreferences.getInstance();
+  //   String firstname = prefs.getString("firstname")?? "no firstname";
+  //   String lastname = prefs.getString("lastname")?? "no lastname";
+  //   String phonenumber = prefs.getString("phonenumber")?? "no phonenumber";
+  //   String email = prefs.getString("email")?? "no email";
+  //
+  //   print("The firstname is $firstname \n thank you");
+  //   print("The lastname is $lastname \n thank you");
+  //   print("The phonenumber is $phonenumber \n thank you");
+  //   print("The email is $email \n thank you");
+  // }
+}
