@@ -1,11 +1,14 @@
+import 'dart:async';
 import 'dart:convert';
 import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:vixx_app/Config/Color.dart';
 import 'package:vixx_app/Config/Style.dart';
 import 'package:vixx_app/Screen/Account/account_info_page.dart';
 import 'package:vixx_app/Screen/HomeScreen/bottom_nav.dart';
 import 'package:http/http.dart' as http;
 import 'package:vixx_app/Screen/Pages/transaction_history.dart';
+import 'package:vixx_app/Screen/Pages/transaction_page.dart';
 
 class DepositPage extends StatefulWidget{
   @override
@@ -13,7 +16,8 @@ class DepositPage extends StatefulWidget{
 }
 
 class DepositPageState extends State <DepositPage>{
-  String amount;
+  Timer? timer;
+  String? amount;
   TextEditingController amountController = TextEditingController();
   GlobalKey<FormState> globalFormKey = GlobalKey<FormState>();
 
@@ -65,7 +69,7 @@ class DepositPageState extends State <DepositPage>{
                   mainAxisAlignment:  MainAxisAlignment.spaceBetween,
                   children: [
                     GestureDetector(
-                      child: Icon(Icons.arrow_back_ios_sharp,size: 20,),
+                      child: Icon(Icons.arrow_back_sharp, size: 25,),
                       onTap: (){
                         Navigator.push(context,MaterialPageRoute(builder: (context) => BottomNavBar()));
                       },
@@ -124,14 +128,14 @@ class DepositPageState extends State <DepositPage>{
                                                 crossAxisAlignment: CrossAxisAlignment.end,
                                                 children: [
                                                   Text("USDT ${totalBal}",style: setStyleContent(context,Colors.black,25,FontWeight.bold)),
-                                                  Text("Available balance",style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.w300)),
+                                                  Text("Available balance",style: setStyleContent(context,ColorConstant.primaryColor,15,FontWeight.w300)),
                                                   SizedBox(height: 15,),
                                                   Column(
                                                       crossAxisAlignment: CrossAxisAlignment.start,
                                                       children: [
                                                         Text('Amount',style: setStyleContent(context,Colors.black,12,FontWeight.bold)),
                                                         TextFormField(
-                                                          keyboardType: TextInputType.text,
+                                                          keyboardType: TextInputType.number,
                                                           controller: amountController,
                                                           onSaved: (val) {
                                                             amount = val;
@@ -160,106 +164,20 @@ class DepositPageState extends State <DepositPage>{
                                                       ]
                                                   ),
                                                   SizedBox(height: 10,),
-                                                  Center(
-                                                    child:  Text("Deposit funds as low as 20USDT",style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.w300)),
-                                                  ),
-                                                  SizedBox(height: 10,),
                                                   GestureDetector(
                                                     child:  Center(
                                                         child: Container(
                                                           height: 40,
                                                           width: 280,
                                                           decoration: BoxDecoration(
-                                                              borderRadius: BorderRadius.circular(8), color: ColorConstant.greenColor),
+                                                              borderRadius: BorderRadius.circular(8), color: ColorConstant.primaryColor),
                                                           child: Center(
                                                             child: Text("Proceed", style: setStyleContent(context, ColorConstant.slightWhiteColor, 12, FontWeight.bold),),
                                                           ),
                                                         )
                                                     ),
                                                     onTap: (){
-                                                      showModalBottomSheet(
-                                                          shape: RoundedRectangleBorder(
-                                                            borderRadius: BorderRadius.only( topLeft: Radius.circular(25),
-                                                                topRight: Radius.circular(25)),
-                                                          ),
-                                                          context: context,
-                                                          isScrollControlled: true,
-                                                          builder: (builder){
-                                                            return Padding(
-                                                              padding:  MediaQuery.of(context).viewInsets,
-                                                              child: Container(
-                                                                height: 300,
-                                                                padding: EdgeInsets.all(20),
-                                                                child: Column(
-                                                                  children: [
-                                                                    SizedBox(height: 10,),
-                                                                    Text("You are about to make a Deposit of",style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.w300)),
-                                                                    SizedBox(height: 10,),
-                                                                    Text("$amount",style: setStyleContent(context,Colors.black,20,FontWeight.bold)),
-                                                                    SizedBox(height: 10,),
-                                                                    Text("To your wallet right now",style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.w300)),
-                                                                    SizedBox(height: 20,),
-                                                                    GestureDetector(
-                                                                      child:  Center(
-                                                                          child: Container(
-                                                                            height: 40,
-                                                                            width: 280,
-                                                                            decoration: BoxDecoration(
-                                                                                borderRadius: BorderRadius.circular(8), color: ColorConstant.greenColor),
-                                                                            child: Center(
-                                                                              child: Text("Confirm", style: setStyleContent(context, ColorConstant.slightWhiteColor, 12, FontWeight.bold),),
-                                                                            ),
-                                                                          )
-                                                                      ),
-                                                                      onTap: ()async{
-                                                                        showDialog(
-                                                                            context: context,
-                                                                            builder: (BuildContext context) {
-                                                                              return Center(child: CircularProgressIndicator(),);
-                                                                            });
-                                                                        await depositAction();
-
-                                                                        // If the form is valid, display a snackbar. In the real world,
-                                                                        // you'd often call a server or save the information in a database.
-                                                                        ScaffoldMessenger.of(context).showSnackBar(
-                                                                          const SnackBar(content: Text('Processing Data'), duration: const Duration(seconds: 2),),
-                                                                        );
-
-                                                                        deposito(amountController.text);
-                                                                        // showModalBottomSheet(
-                                                                        //     shape: RoundedRectangleBorder(
-                                                                        //       borderRadius: BorderRadius.only( topLeft: Radius.circular(25),
-                                                                        //           topRight: Radius.circular(25)),
-                                                                        //     ),
-                                                                        //     context: context,
-                                                                        //     builder: (builder){
-                                                                        //       return Padding(
-                                                                        //         padding:  MediaQuery.of(context).viewInsets,
-                                                                        //         child: Container(
-                                                                        //           height: 300,
-                                                                        //           padding: EdgeInsets.all(20),
-                                                                        //           child: Column(
-                                                                        //             children: [
-                                                                        //               SizedBox(height: 10,),
-                                                                        //               Image.asset("assets/images/success.png",height: 100,),
-                                                                        //               SizedBox(height: 10,),
-                                                                        //               Text("Transaction\n  completed", style: setStyleContent(context, ColorConstant.greenColor, 18, FontWeight.bold),)
-                                                                        //             ],
-                                                                        //           ),
-                                                                        //         ),
-                                                                        //
-                                                                        //       );
-                                                                        //     }
-                                                                        // );
-                                                                      },
-                                                                    ),
-                                                                  ],
-                                                                ),
-                                                              ),
-
-                                                            );
-                                                          }
-                                                      );
+                                                      depowarning(context, amountController.text);
                                                     },
                                                   ),
                                                 ],
@@ -282,7 +200,7 @@ class DepositPageState extends State <DepositPage>{
                   children: [
                     Text("Transaction history", style: setStyleContent(context,ColorConstant.black,18,FontWeight.w600)),
                     GestureDetector(
-                      child:  Icon(Icons.arrow_forward_ios_sharp, size: 20,),
+                      child:   Text("View all", style: setStyleContent(context,ColorConstant.black,13,FontWeight.w600)),
                       onTap: (){
                         Navigator.push(context,MaterialPageRoute(builder: (context) => TransactionHistoryPage()));
                       },
@@ -291,432 +209,74 @@ class DepositPageState extends State <DepositPage>{
                   ],
                 ),
                 SizedBox(height: 15,),
-                Column(
-                  children: [
-                    GestureDetector(
-                      child:  Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.pinkColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$3,019.00", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("2 days ago", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    SizedBox(height: 10,),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Failed", style: setStyleContent(context,ColorConstant.pinkColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$7,038", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Sep 27, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child:  Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.pinkColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Failed", style: setStyleContent(context,ColorConstant.pinkColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 01, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Failed", style: setStyleContent(context,ColorConstant.pinkColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 01, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child:  Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.pinkColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 01, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    Divider(),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 15, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 15, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child:  Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.pinkColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 01, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    Divider(),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 15, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child:  Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.pinkColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 01, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 15, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("Deposit", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 15, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                  ],
-                )
+                DepositTransactionPage()
               ],
             ),
           ),
         ),
       ),
+    );
+  }
+
+
+
+  depowarning(BuildContext context, String text) async {
+    String teamName = '';
+    return showDialog(
+      context: context,
+      barrierDismissible: false, // dialog is dismissible with a tap on the barrier
+      builder: (BuildContext context) {
+        return AlertDialog(
+            content: SingleChildScrollView(
+                child:  Padding(
+                  padding:  MediaQuery.of(context).viewInsets,
+                  child: Container(
+                    height: 200,
+                    padding: EdgeInsets.all(20),
+                    child: Column(
+                      children: [
+                        SizedBox(height: 10,),
+                        Text("You are about to make a\ndeposit of",style: setStyleContent(context,ColorConstant.black,14,FontWeight.w500), textAlign: TextAlign.center),
+                        SizedBox(height: 10,),
+                        Text("usdt ${amountController.text}",style: setStyleContent(context,Colors.black,20,FontWeight.bold)),
+                        SizedBox(height: 5,),
+                        GestureDetector(
+                          child:  Center(
+                              child: Container(
+                                height: 35,
+                                width: 280,
+                                decoration: BoxDecoration(
+                                    borderRadius: BorderRadius.circular(8), color: ColorConstant.primaryColor),
+                                child: Center(
+                                  child: Text("Confirm", style: setStyleContent(context, ColorConstant.slightWhiteColor, 12, FontWeight.bold),),
+                                ),
+                              )
+                          ),
+                          onTap: () async{
+                            showDialog(
+                                context: context,
+                                builder: (BuildContext context) {
+                                  return Center(child: CircularProgressIndicator(
+                                    color: ColorConstant.primaryColor,
+                                  ),);
+                                });
+                            await depositAction();
+
+                            // If the form is valid, display a snackbar. In the real world,
+                            // you'd often call a server or save the information in a database.
+                            ScaffoldMessenger.of(context).showSnackBar(
+                              const SnackBar(content: Text('Processing Data'), duration: const Duration(seconds: 2),),
+                            );
+                            deposito(amountController.text);
+                          },
+                        ),
+                      ],
+                    ),
+                  ),
+                )
+            )
+        );
+      },
     );
   }
 
@@ -733,12 +293,14 @@ class DepositPageState extends State <DepositPage>{
     };
     print(data);
     String body = json.encode(data);
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    String? apikey = prefs.getString("apikey");
     String Url = "https://thevix.club/apimaster.php";
     var response = await http.post(Uri.parse(Url),
         headers: {
           "Content-Type": "application/json",
           "Operation": "summary",
-          "Authorization": "hthwijlewivcbeusnwjx6yqldbdi"
+          "Authorization": "$apikey"
         },
 
 
@@ -764,12 +326,14 @@ class DepositPageState extends State <DepositPage>{
     };
     print(data);
     String body = json.encode(data);
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    String? apikey = prefs.getString("apikey");
     String Url = "https://thevix.club/apimaster.php";
     var response = await http.post(Uri.parse(Url),
         headers: {
           "Content-Type": "application/json",
           "Operation": "deposit",
-          "Authorization": "hthwijlewivcbeusnwjx6yqldbdi"
+          "Authorization": "$apikey"
         },
 
 
