@@ -1,11 +1,15 @@
+import 'dart:async';
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:vixx_app/Config/Color.dart';
 import 'package:vixx_app/Config/Style.dart';
 import 'package:vixx_app/Screen/Account/account_info_page.dart';
 import 'package:vixx_app/Screen/HomeScreen/bottom_nav.dart';
 import 'package:http/http.dart' as http;
+import 'package:vixx_app/Screen/Pages/invest_trans_page.dart';
 import 'package:vixx_app/Screen/Pages/transaction_history.dart';
 
 class InvestPage extends StatefulWidget{
@@ -14,8 +18,9 @@ class InvestPage extends StatefulWidget{
 }
 
 class InvestPageState extends State <InvestPage>{
+  Timer? timer;
 
-  String amount;
+  String? amount;
   TextEditingController amountController = TextEditingController();
   GlobalKey<FormState> globalFormKey = GlobalKey<FormState>();
 
@@ -63,7 +68,7 @@ class InvestPageState extends State <InvestPage>{
                   mainAxisAlignment:  MainAxisAlignment.spaceBetween,
                   children: [
                     GestureDetector(
-                      child: Icon(Icons.arrow_back_ios_sharp, size: 20,),
+                      child: Icon(Icons.arrow_back_sharp, size: 25,),
                       onTap: (){
                         Navigator.push(context,MaterialPageRoute(builder: (context) => BottomNavBar()));
                       },
@@ -122,14 +127,14 @@ class InvestPageState extends State <InvestPage>{
                                             crossAxisAlignment: CrossAxisAlignment.end,
                                             children: [
                                               Text("USDT ${totalBal}",style: setStyleContent(context,Colors.black,25,FontWeight.bold)),
-                                              Text("Available balance",style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.w300)),
+                                              Text("Available balance",style: setStyleContent(context,ColorConstant.primaryColor,15,FontWeight.w300)),
                                               SizedBox(height: 15,),
                                               Column(
                                                   crossAxisAlignment: CrossAxisAlignment.start,
                                                   children: [
                                                     Text('Amount',style: setStyleContent(context,Colors.black,12,FontWeight.bold)),
                                                     TextFormField(
-                                                      keyboardType: TextInputType.text,
+                                                      keyboardType: TextInputType.number,
                                                       controller: amountController,
                                                       onSaved: (val) {
                                                         amount = val;
@@ -152,105 +157,20 @@ class InvestPageState extends State <InvestPage>{
                                                   ]
                                               ),
                                               SizedBox(height: 10,),
-                                              Center(
-                                                child:  Text("Get a return of 17% in 12 months",style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.w300)),
-                                              ),
-                                              SizedBox(height: 10,),
                                               GestureDetector(
                                                 child:  Center(
                                                     child: Container(
                                                       height: 40,
                                                       width: 280,
                                                       decoration: BoxDecoration(
-                                                          borderRadius: BorderRadius.circular(8), color: ColorConstant.greenColor),
+                                                          borderRadius: BorderRadius.circular(8), color: ColorConstant.primaryColor),
                                                       child: Center(
                                                         child: Text("Proceed", style: setStyleContent(context, ColorConstant.slightWhiteColor, 12, FontWeight.bold),),
                                                       ),
                                                     )
                                                 ),
-                                                onTap: ()async{
-                                                  showModalBottomSheet(
-                                                      shape: RoundedRectangleBorder(
-                                                        borderRadius: BorderRadius.only( topLeft: Radius.circular(25),
-                                                            topRight: Radius.circular(25)),
-                                                      ),
-                                                      context: context,
-                                                      isScrollControlled: true,
-                                                      builder: (builder){
-                                                        return Padding(
-                                                          padding:  MediaQuery.of(context).viewInsets,
-                                                          child: Container(
-                                                            height: 300,
-                                                            padding: EdgeInsets.all(20),
-                                                            child: Column(
-                                                              children: [
-                                                                SizedBox(height: 10,),
-                                                                Text("You are about to make an investment of",style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.w300)),
-                                                                SizedBox(height: 10,),
-                                                                Text("$amount",style: setStyleContent(context,Colors.black,20,FontWeight.bold)),
-                                                                SizedBox(height: 10,),
-                                                                Text("To get a return of 17% in 12 months",style: setStyleContent(context,ColorConstant.greenColor,15,FontWeight.w300)),
-                                                                SizedBox(height: 20,),
-                                                                GestureDetector(
-                                                                  child:  Center(
-                                                                      child: Container(
-                                                                        height: 35,
-                                                                        width: 280,
-                                                                        decoration: BoxDecoration(
-                                                                            borderRadius: BorderRadius.circular(8), color: ColorConstant.greenColor),
-                                                                        child: Center(
-                                                                          child: Text("Confirm", style: setStyleContent(context, ColorConstant.slightWhiteColor, 12, FontWeight.bold),),
-                                                                        ),
-                                                                      )
-                                                                  ),
-                                                                  onTap: () async{
-                                                                    showDialog(
-                                                                        context: context,
-                                                                        builder: (BuildContext context) {
-                                                                          return Center(child: CircularProgressIndicator(),);
-                                                                        });
-                                                                    await investAction();
-
-                                                                      // If the form is valid, display a snackbar. In the real world,
-                                                                      // you'd often call a server or save the information in a database.
-                                                                      ScaffoldMessenger.of(context).showSnackBar(
-                                                                        const SnackBar(content: Text('Processing Data'), duration: const Duration(seconds: 2),),
-                                                                      );
-                                                                    investo(amountController.text);
-                                                                    // showModalBottomSheet(
-                                                                    //     shape: RoundedRectangleBorder(
-                                                                    //       borderRadius: BorderRadius.only( topLeft: Radius.circular(25),
-                                                                    //           topRight: Radius.circular(25)),
-                                                                    //     ),
-                                                                    //     context: context,
-                                                                    //     builder: (builder){
-                                                                    //       return Padding(
-                                                                    //         padding:  MediaQuery.of(context).viewInsets,
-                                                                    //         child: Container(
-                                                                    //           height: 300,
-                                                                    //           padding: EdgeInsets.all(20),
-                                                                    //           child: Column(
-                                                                    //             children: [
-                                                                    //               SizedBox(height: 10,),
-                                                                    //               Image.asset("assets/images/success.png",height: 100,),
-                                                                    //               SizedBox(height: 10,),
-                                                                    //               Text("Transaction\n  completed", style: setStyleContent(context, ColorConstant.greenColor, 18, FontWeight.bold),)
-                                                                    //             ],
-                                                                    //           ),
-                                                                    //         ),
-                                                                    //
-                                                                    //       );
-                                                                    //     }
-                                                                    // );
-                                                                  },
-                                                                ),
-                                                              ],
-                                                            ),
-                                                          ),
-
-                                                        );
-                                                      }
-                                                  );
+                                                onTap: (){
+                                                  warning(context, amountController.text);
                                                 },
                                               ),
                                             ],
@@ -273,7 +193,7 @@ class InvestPageState extends State <InvestPage>{
                   children: [
                     Text("Transaction history", style: setStyleContent(context,ColorConstant.black,18,FontWeight.w600)),
                    GestureDetector(
-                     child:  Icon(Icons.arrow_forward_ios_sharp, size: 20,),
+                     child:  Text("View all", style: setStyleContent(context,ColorConstant.black,13,FontWeight.w600)),
                      onTap: (){
                        Navigator.push(context,MaterialPageRoute(builder: (context) => TransactionHistoryPage()));
                      },
@@ -282,219 +202,7 @@ class InvestPageState extends State <InvestPage>{
                       ],
                     ),
                 SizedBox(height: 15,),
-                Column(
-                  children: [
-                    GestureDetector(
-                      child:  Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.pinkColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("USDT Withdrawal", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$3,019.00", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("2 days ago", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    SizedBox(height: 10,),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("USDT Withdrawal", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Failed", style: setStyleContent(context,ColorConstant.pinkColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$7,038", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Sep 27, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child:  Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.pinkColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("BTC Withdrawal", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Failed", style: setStyleContent(context,ColorConstant.pinkColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 01, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("BTC Withdrawal", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Failed", style: setStyleContent(context,ColorConstant.pinkColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 01, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child:  Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.pinkColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("BTC Withdrawal", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 01, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: 10,),
-                    Divider(),
-                    GestureDetector(
-                      child: Row(
-                        children: [
-                          Container(
-                            padding: EdgeInsets.all(10),
-                            height: 40,
-                            width: 40,
-                            decoration: BoxDecoration(
-                                borderRadius: BorderRadius.circular(8),
-                                color: ColorConstant.greenColor
-                            ),
-                            child: Center(
-                              child: Image.asset("assets/images/wallet_icon.png", height: 25,),
-                            ),
-                          ),
-                          SizedBox(width: 12,),
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text("USDT Withdrawal", style: setStyleContent(context,ColorConstant.black,12,FontWeight.w400)),
-                              Text("Successful", style: setStyleContent(context,ColorConstant.greenColor,10,FontWeight.w400)),
-                            ],
-                          ),
-                          Spacer(),
-                          Column(
-                            children: [
-                              Text("\$60,337", style: setStyleContent(context,ColorConstant.black,12,FontWeight.bold)),
-                              Text("Oct 15, 2019", style: setStyleContent(context,ColorConstant.black,10,FontWeight.w300))
-                            ],
-                          )
-                        ],
-                      ),
-                    ),
-                  ],
-                )
+                InvestTransactionPage()
               ],
             ),
           ),
@@ -502,6 +210,68 @@ class InvestPageState extends State <InvestPage>{
       ),
     );
   }
+
+  warning(BuildContext context, String text) async {
+    String teamName = '';
+    return showDialog(
+      context: context,
+      barrierDismissible: false, // dialog is dismissible with a tap on the barrier
+      builder: (BuildContext context) {
+        return AlertDialog(
+            content: SingleChildScrollView(
+              child:  Padding(
+                padding:  MediaQuery.of(context).viewInsets,
+                child: Container(
+                  height: 200,
+                  padding: EdgeInsets.all(20),
+                  child: Column(
+                    children: [
+                      SizedBox(height: 10,),
+                      Text("You are about to make an\ninvestment of",style: setStyleContent(context,ColorConstant.black,14,FontWeight.w500), textAlign: TextAlign.center),
+                      SizedBox(height: 10,),
+                      Text("usdt ${amountController.text}",style: setStyleContent(context,Colors.black,20,FontWeight.bold)),
+                      SizedBox(height: 5,),
+                      GestureDetector(
+                        child:  Center(
+                            child: Container(
+                              height: 35,
+                              width: 280,
+                              decoration: BoxDecoration(
+                                  borderRadius: BorderRadius.circular(8), color: ColorConstant.primaryColor),
+                              child: Center(
+                                child: Text("Confirm", style: setStyleContent(context, ColorConstant.slightWhiteColor, 12, FontWeight.bold),),
+                              ),
+                            )
+                        ),
+                        onTap: () async{
+                          showDialog(
+                              context: context,
+                              builder: (BuildContext context) {
+                                return Center(child: CircularProgressIndicator(
+                                  color: ColorConstant.primaryColor,
+                                ),);
+                              });
+                          await investAction();
+
+                          // If the form is valid, display a snackbar. In the real world,
+                          // you'd often call a server or save the information in a database.
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            const SnackBar(content: Text('Processing Data'), duration: const Duration(seconds: 2),),
+                          );
+                          investo(amountController.text);
+                        },
+                      ),
+                    ],
+                  ),
+                ),
+              )
+            )
+        );
+      },
+    );
+  }
+
+
 
   Future<bool> investAction() async {
     //replace the below line of code with your login request
@@ -515,12 +285,14 @@ class InvestPageState extends State <InvestPage>{
     };
     print(data);
     String body = json.encode(data);
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    String? apikey = prefs.getString("apikey");
     String Url = "https://thevix.club/apimaster.php";
     var response = await http.post(Uri.parse(Url),
         headers: {
           "Content-Type": "application/json",
           "Operation": "summary",
-          "Authorization": "hthwijlewivcbeusnwjx6yqldbdi"
+          "Authorization": "$apikey"
         },
 
 
@@ -546,12 +318,14 @@ class InvestPageState extends State <InvestPage>{
     };
     print(data);
     String body = json.encode(data);
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    String? apikey = prefs.getString("apikey");
     String Url = "https://thevix.club/apimaster.php";
     var response = await http.post(Uri.parse(Url),
         headers: {
           "Content-Type": "application/json",
           "Operation": "invest",
-          "Authorization": "hthwijlewivcbeusnwjx6yqldbdi"
+          "Authorization": "$apikey"
         },
 
 
