@@ -8,9 +8,14 @@ import 'package:vixx_app/Screen/Pages/deposit_page.dart';
 import 'package:vixx_app/Screen/Pages/invest_page.dart';
 import 'package:vixx_app/Screen/Pages/transaction_history.dart';
 import 'package:vixx_app/Screen/Pages/withdrawal_page.dart';
+import 'package:vixx_app/Screen/Teams/team_list.dart';
 import 'package:vixx_app/Screen/Teams/team_page.dart';
 
 class BottomNavBar extends StatefulWidget {
+
+  BottomNavBar({this.api_key});
+  final api_key;
+
   @override
   BottomNavBarState createState() => new BottomNavBarState();
 }
@@ -26,6 +31,8 @@ class BottomNavBarState extends State<BottomNavBar> {
   int _activePage = 0;
   GlobalKey<CurvedNavigationBarState> _bottomNavigationKey = GlobalKey();
 
+  static String? get username => null;
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -34,11 +41,11 @@ class BottomNavBarState extends State<BottomNavBar> {
           index: 0,
           height: 60.0,
           items: <Widget>[
-            Icon(Icons.home_rounded, size: 30, color: ColorConstant.primaryColor,),
-            Icon(Icons.trending_up_sharp, size: 30 , color: ColorConstant.primaryColor,),
-            Icon(Icons.add, size: 30 , color: ColorConstant.primaryColor,),
-            Icon(Icons.account_balance_wallet_rounded, size: 30 , color: ColorConstant.primaryColor,),
-            Icon(Icons.people_alt_rounded,  size: 30 , color: ColorConstant.primaryColor, ),
+            Image.asset("assets/images/hmm.png", height: 25,color: ColorConstant.primaryColor,),
+            Image.asset("assets/images/in_icon.png", height: 30,color: ColorConstant.primaryColor,),
+            Image.asset("assets/images/wallet_icon.png", height: 30,color: ColorConstant.primaryColor,),
+            Image.asset("assets/images/with_icon.png", height: 30,color: ColorConstant.primaryColor,),
+            Image.asset("assets/images/grp_inc.png", height: 50,color: ColorConstant.primaryColor,),
           ],
           color: Colors.white,
           buttonBackgroundColor: Colors.white,
